Factor PythonArrowInput out as a symmetry to PythonArrowOutput
